+++ src/Benchmark/Program.cs
@@ -3,6 +3,7 @@ using System.Text;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
 using JustEat.StatsD;
+using JustEat.StatsD.V2;
 
 namespace Benchmark
 {
@@ -24,24 +25,24 @@ namespace Benchmark
 
         private static readonly StatsDMessageFormatter FormatterOld = new StatsDMessageFormatter("hello.world");
 
-        //[Benchmark(Baseline = true)]
+        [Benchmark(Baseline = true)]
         public void Original()
         {
             Encoding.UTF8.GetBytes(FormatterOld.Gauge(255.1, "some.neat.bucket"));
         }
 
-        [Benchmark(Baseline = true)]
+        [Benchmark]
         public void HeapBuffer()
         {
             Formatter.TryFormat(StatsDMessage.Gauge(255.1, "some.neat.bucket"), 1, Buffer, out var written);
         }
 
-        //[Benchmark]
-        //public void StackBuffer()
-        //{
-        //    var statsDMessage = StatsDMessage.Counter(1, "some.neat.bucket");
-        //    Span<byte> buffer = stackalloc byte[Formatter.GetBufferSize(statsDMessage)];
-        //    Formatter.TryFormat(statsDMessage, 1, buffer, out var written);
-        //}
+        [Benchmark]
+        public void StackBuffer()
+        {
+            var statsDMessage = StatsDMessage.Counter(1, "some.neat.bucket");
+            Span<byte> buffer = stackalloc byte[Formatter.GetBufferSize(statsDMessage)];
+            Formatter.TryFormat(statsDMessage, 1, buffer, out var written);
+        }
     }
 }
