+++ src/JustEat.StatsD/Properties/AssemblyInfo.cs
@@ -1,4 +1,8 @@
-﻿using System.Runtime.InteropServices;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 [assembly: ComVisible(false)]
 [assembly: Guid("8f4ff09e-4130-4872-a50f-b290e9ccb04b")]
+
+[assembly:InternalsVisibleTo("JustEat.StatsD.Tests")]
+[assembly:InternalsVisibleTo("Benchmark")]
