+++ src/JustEat.StatsD/V2/IStatsDTransportV2.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace JustEat.StatsD.V2
+{
+    public interface IStatsDTransportV2
+    {
+        void Send(ArraySegment<byte> metric);
+    }
+}
